refactor: split api.ts into domain-scoped API modules (#833)
Replace the monolithic frontend/src/api.ts (1040 lines, every domain
mixed together) with frontend/src/api/ sub-modules (core, articles,
sources, users, scheduler, ask, stats, briefings, auth, admin,
notifications, settings, shares, quiz, onboarding, personalization,
watchlists) plus a barrel index.ts so existing `from '@/api'` imports
keep working unchanged. workflowApi.ts/readingListApi.ts/tagsApi.ts
now import shared helpers from ./core directly instead of the barrel.

Closes #833
