--- frontend/src/api.ts
@@ -1,1040 +1,0 @@
-import type {
-  AdminAnalytics,
-  AdminAiQuality,
-  AgentActionPlanResponse,
-  AgentActionRun,
-  AiMemory,
-  AiWatchlist,
-  AiWatchlistMatch,
-  AiWatchlistNudge,
-  AnalyticsSettings,
-  AnalyticsSettingsUpdate,
-  Article,
-  ArticleCountsResult,
-  ArticleHighlight,
-  ArticleStatus,
-  ArticlesOverTimePoint,
-  AskResponse,
-  Briefing,
-  BriefingCreateResponse,
-  BriefingLatestResponse,
-  CategoryMixPoint,
-  IngestedVsHandledPoint,
-  McpToken,
-  NotificationSettings,
-  NotificationSettingsUpdate,
-  OnboardingInterest,
-  OnboardingSourceRecommendation,
-  OnboardingStatus,
-  OpmlImportResult,
-  PersonalizationNudge,
-  PushSubscribeRequest,
-  Quiz,
-  QuizCandidate,
-  QuizHistoryItem,
-  QuizResult,
-  Achievement,
-  ReadingDna,
-  ReadingGoal,
-  ReadingStreak,
-  RecommendationPreferences,
-  ReceivedShare,
-  SentShare,
-  ShareDetail,
-  ShareAnnotation,
-  ShareMessage,
-  SaveOnboardingInterestsRequest,
-  ShareableUser,
-  Source,
-  SourceCleanupSuggestion,
-  SourceHealth,
-  SourceQualityRow,
-  SourceVolumePoint,
-  StatsOverview,
-  Summary,
-  TriageMetrics,
-  IngestRunPage,
-  IngestRunSource,
-  TopicMapResponse,
-  WordCloudResponse,
-  EmbeddingMapResponse,
-  KnowledgeGraphResponse,
-  User,
-  WeeklyRecap,
-} from './types';
-
-export async function readErrorMessage(response: Response): Promise<string> {
-  try {
-    const body: unknown = await (response.json() as Promise<unknown>);
-    if (body && typeof body === 'object') {
-      const { detail, message } = body as Record<string, unknown>;
-      if (typeof detail === 'string' && detail) return detail;
-      if (Array.isArray(detail) && detail.length > 0) {
-        const msgs = detail
-          .map((d) => (d && typeof d === 'object' ? (d as Record<string, unknown>).msg : null))
-          .filter((m): m is string => typeof m === 'string');
-        if (msgs.length > 0) return msgs.join('; ');
-      }
-      if (typeof message === 'string' && message) return message;
-    }
-  } catch {
-    // non-JSON body — fall through
-  }
-  return `${response.status} ${response.statusText}`;
-}
-
-export class HttpError extends Error {
-  constructor(
-    public readonly status: number,
-    message: string
-  ) {
-    super(message);
-    this.name = 'HttpError';
-  }
-}
-
-export async function requestJson<T>(url: string, init?: RequestInit): Promise<T> {
-  const response = await fetch(url, {
-    headers: { 'Content-Type': 'application/json', ...(init?.headers ?? {}) },
-    credentials: 'same-origin',
-    ...init,
-  });
-  if (!response.ok) {
-    throw new HttpError(response.status, await readErrorMessage(response));
-  }
-  return response.json() as Promise<T>;
-}
-
-export async function fetchArticles(
-  status?: ArticleStatus,
-  category?: string,
-  offset = 0,
-  limit = 100
-): Promise<Article[]> {
-  const params = new URLSearchParams();
-  if (status) params.set('status', status);
-  if (category) params.set('category', category);
-  if (offset > 0) params.set('offset', String(offset));
-  if (limit !== 100) params.set('limit', String(limit));
-  const suffix = params.size ? `?${params}` : '';
-  const data = await requestJson<{ items: Article[] }>(`/api/articles${suffix}`);
-  return data.items;
-}
-
-export async function searchArticles(q: string, limit = 50): Promise<Article[]> {
-  const params = new URLSearchParams({ q, limit: String(limit) });
-  const data = await requestJson<{ items: Article[] }>(`/api/search?${params}`);
-  return data.items;
-}
-
-export async function fetchArticle(id: number | string): Promise<Article> {
-  return requestJson<Article>(`/api/articles/${id}`);
-}
-
-export async function fetchArticleBody(id: number | string): Promise<Article> {
-  return requestJson<Article>(`/api/articles/${id}/body`);
-}
-
-export async function saveSharedUrl(payload: {
-  url: string;
-  title?: string | null;
-  text?: string | null;
-}): Promise<Article> {
-  return requestJson<Article>('/api/articles/save-url', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function fetchArticleHighlights(id: number | string): Promise<ArticleHighlight[]> {
-  const data = await requestJson<{ items: ArticleHighlight[] }>(`/api/articles/${id}/highlights`);
-  return data.items;
-}
-
-export async function createArticleHighlight(
-  id: number | string,
-  payload: { highlighted_text: string; offset_chars?: number; note?: string | null }
-): Promise<ArticleHighlight> {
-  return requestJson<ArticleHighlight>(`/api/articles/${id}/highlights`, {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function deleteArticleHighlight(
-  articleId: number | string,
-  highlightId: number | string
-): Promise<{ ok: boolean }> {
-  return requestJson<{ ok: boolean }>(`/api/articles/${articleId}/highlights/${highlightId}`, {
-    method: 'DELETE',
-  });
-}
-
-export async function fetchSharedArticle(shareId: number | string): Promise<Article> {
-  return requestJson<Article>(`/api/shares/${shareId}/article`);
-}
-
-export async function fetchSharedArticleBody(shareId: number | string): Promise<Article> {
-  return requestJson<Article>(`/api/shares/${shareId}/article/body`, { method: 'POST' });
-}
-
-export async function fetchArticleAudioUrl(id: number | string): Promise<string> {
-  const response = await fetch(`/api/articles/${id}/audio`, {
-    method: 'POST',
-    credentials: 'same-origin',
-  });
-  if (!response.ok) {
-    throw new Error(await readErrorMessage(response));
-  }
-  const blob = await response.blob();
-  return URL.createObjectURL(blob);
-}
-
-export async function fetchArticleInsights(id: number | string): Promise<string[]> {
-  const data = await requestJson<{ bullets: string[] }>(`/api/articles/${id}/insights`);
-  return data.bullets;
-}
-
-export interface PerspectiveAnalysis {
-  verified_facts: string[];
-  omissions: string[];
-  alternative_perspectives: string[];
-}
-
-export async function fetchArticlePerspectives(id: number | string): Promise<PerspectiveAnalysis> {
-  return requestJson<PerspectiveAnalysis>(`/api/articles/${id}/perspectives`);
-}
-
-export async function fetchSources(): Promise<Source[]> {
-  const data = await requestJson<{ items: Source[] }>('/api/sources');
-  return data.items;
-}
-
-export async function fetchSourceHealth(): Promise<SourceHealth[]> {
-  const data = await requestJson<{ items: SourceHealth[] }>('/api/sources/health');
-  return data.items;
-}
-
-export async function fetchSourceCleanupSuggestions(): Promise<SourceCleanupSuggestion[]> {
-  const data = await requestJson<{ items: SourceCleanupSuggestion[] }>(
-    '/api/sources/cleanup-suggestions'
-  );
-  return data.items;
-}
-
-export async function applySourceCleanup(sourceSlugs: string[]): Promise<{
-  updated: string[];
-  skipped: string[];
-}> {
-  return requestJson('/api/sources/cleanup', {
-    method: 'POST',
-    body: JSON.stringify({ source_slugs: sourceSlugs }),
-  });
-}
-
-export interface CreateSourcePayload {
-  url: string;
-  name: string;
-  category?: string;
-  slug?: string;
-  kind?: string;
-}
-
-export async function createSource(payload: CreateSourcePayload): Promise<Source> {
-  return requestJson<Source>('/api/sources', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export interface PreviewSourcePayload {
-  url: string;
-  kind?: string;
-}
-
-export interface SourcePreviewItem {
-  title: string;
-  url: string;
-  date: string | null;
-}
-
-export interface SourcePreviewResult {
-  kind: string;
-  entry_count: number;
-  items: SourcePreviewItem[];
-}
-
-export async function previewSource(payload: PreviewSourcePayload): Promise<SourcePreviewResult> {
-  return requestJson<SourcePreviewResult>('/api/sources/preview', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function deleteSource(slug: string): Promise<{ status: string }> {
-  return requestJson<{ status: string }>(`/api/sources/${slug}`, { method: 'DELETE' });
-}
-
-export async function fetchSummary(): Promise<Summary> {
-  return requestJson<Summary>('/api/summary');
-}
-
-export async function fetchReadingDna(): Promise<ReadingDna> {
-  return requestJson<ReadingDna>('/api/users/me/reading-dna');
-}
-
-export async function fetchReadingStreak(): Promise<ReadingStreak> {
-  return requestJson<ReadingStreak>('/api/users/me/streak');
-}
-
-export async function fetchAchievements(): Promise<Achievement[]> {
-  const data = await requestJson<{ items: Achievement[] }>('/api/users/me/achievements');
-  return data.items;
-}
-
-export async function fetchRecaps(): Promise<WeeklyRecap[]> {
-  const data = await requestJson<{ items: WeeklyRecap[] }>('/api/recaps');
-  return data.items;
-}
-
-export async function fetchLatestRecap(): Promise<WeeklyRecap | null> {
-  try {
-    return await requestJson<WeeklyRecap>('/api/recaps/latest');
-  } catch (err) {
-    if (err instanceof HttpError && err.status === 404) return null;
-    throw err;
-  }
-}
-
-export async function fetchRecommendationPreferences(): Promise<RecommendationPreferences> {
-  return requestJson<RecommendationPreferences>('/api/users/me/recommendation-preferences');
-}
-
-export async function saveRecommendationPreferences(
-  preferences: Partial<RecommendationPreferences>
-): Promise<RecommendationPreferences> {
-  return requestJson<RecommendationPreferences>('/api/users/me/recommendation-preferences', {
-    method: 'PATCH',
-    body: JSON.stringify(preferences),
-  });
-}
-
-export async function fetchAiMemories(): Promise<AiMemory[]> {
-  const data = await requestJson<{ items: AiMemory[] }>('/api/users/me/ai-memories');
-  return data.items;
-}
-
-export async function createAiMemory(content: string): Promise<AiMemory> {
-  return requestJson<AiMemory>('/api/users/me/ai-memories', {
-    method: 'POST',
-    body: JSON.stringify({ content }),
-  });
-}
-
-export async function updateAiMemory(
-  memoryId: number,
-  payload: Partial<Pick<AiMemory, 'content' | 'memory_type' | 'confidence' | 'active'>>
-): Promise<AiMemory> {
-  return requestJson<AiMemory>(`/api/users/me/ai-memories/${memoryId}`, {
-    method: 'PATCH',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function deactivateAiMemory(memoryId: number): Promise<AiMemory> {
-  return requestJson<AiMemory>(`/api/users/me/ai-memories/${memoryId}`, {
-    method: 'DELETE',
-  });
-}
-
-export async function learnAiMemoriesFromReading(): Promise<AiMemory[]> {
-  const data = await requestJson<{ items: AiMemory[] }>(
-    '/api/users/me/ai-memories/learn-from-reading',
-    { method: 'POST' }
-  );
-  return data.items;
-}
-
-export async function fetchMcpTokens(): Promise<{ items: McpToken[]; enabled: boolean }> {
-  return requestJson('/api/users/me/mcp-tokens');
-}
-
-export async function createMcpToken(name: string): Promise<McpToken> {
-  return requestJson<McpToken>('/api/users/me/mcp-tokens', {
-    method: 'POST',
-    body: JSON.stringify({ name }),
-  });
-}
-
-export async function revokeMcpToken(tokenId: number): Promise<McpToken> {
-  return requestJson<McpToken>(`/api/users/me/mcp-tokens/${tokenId}`, {
-    method: 'DELETE',
-  });
-}
-
-export async function ingestNow(): Promise<{
-  inserted: number;
-  results: Record<string, number>;
-  run_id: number;
-  total_errors: number;
-  failed_sources: string[];
-}> {
-  return requestJson('/api/ingest', { method: 'POST' });
-}
-
-export async function updateArticleStatus(id: number, status: ArticleStatus): Promise<Article> {
-  return requestJson<Article>(`/api/articles/${id}/status`, {
-    method: 'PATCH',
-    body: JSON.stringify({ status }),
-  });
-}
-
-export async function askAI(query: string, includeAll = false): Promise<AskResponse> {
-  return requestJson<AskResponse>('/api/ask', {
-    method: 'POST',
-    body: JSON.stringify({ query, include_all: includeAll }),
-  });
-}
-
-export async function planAgentActions(query: string): Promise<AgentActionPlanResponse> {
-  return requestJson<AgentActionPlanResponse>('/api/agent/actions/plan', {
-    method: 'POST',
-    body: JSON.stringify({ query }),
-  });
-}
-
-export async function approveAgentActionRun(runId: number): Promise<AgentActionRun> {
-  return requestJson<AgentActionRun>(`/api/agent/actions/${runId}/approve`, { method: 'POST' });
-}
-
-export async function cancelAgentActionRun(runId: number): Promise<AgentActionRun> {
-  return requestJson<AgentActionRun>(`/api/agent/actions/${runId}/cancel`, { method: 'POST' });
-}
-
-export async function submitFeedback(
-  traceId: string,
-  helpful: boolean,
-  comment?: string
-): Promise<{ recorded: boolean }> {
-  return requestJson<{ recorded: boolean }>('/api/feedback', {
-    method: 'POST',
-    body: JSON.stringify({ trace_id: traceId, helpful, comment }),
-  });
-}
-
-export async function updateSourceEnabled(slug: string, enabled: boolean): Promise<Source> {
-  return requestJson<Source>(`/api/sources/${slug}/enabled`, {
-    method: 'PATCH',
-    body: JSON.stringify({ enabled }),
-  });
-}
-
-export interface SchedulerStatus {
-  interval_minutes: number;
-  paused: boolean;
-  next_run_at: string | null;
-  interval_ingest_enabled?: boolean;
-  ingest_authority?: 'in_process' | 'external';
-}
-
-export async function fetchSchedulerStatus(): Promise<SchedulerStatus> {
-  return requestJson<SchedulerStatus>('/api/scheduler/status');
-}
-
-export async function setSchedulerInterval(
-  minutes: number
-): Promise<{ interval_minutes: number; next_run_at: string | null }> {
-  return requestJson('/api/scheduler/interval', {
-    method: 'POST',
-    body: JSON.stringify({ minutes }),
-  });
-}
-
-export async function pauseScheduler(): Promise<{ paused: boolean }> {
-  return requestJson('/api/scheduler/pause', { method: 'POST' });
-}
-
-export async function resumeScheduler(): Promise<{ paused: boolean; next_run_at: string | null }> {
-  return requestJson('/api/scheduler/resume', { method: 'POST' });
-}
-
-export interface ScheduledJobRun {
-  id: number;
-  job_name: string;
-  started_at: string;
-  finished_at: string | null;
-  duration_ms: number | null;
-  status: 'success' | 'skipped' | 'failure';
-  message: string | null;
-}
-
-export async function fetchLatestJobRuns(): Promise<ScheduledJobRun[]> {
-  const data = await requestJson<{ items: ScheduledJobRun[] }>('/api/scheduler/job-runs');
-  return data.items;
-}
-
-function statsParams(from: string, to: string): string {
-  return new URLSearchParams({ from, to }).toString();
-}
-
-export async function fetchStatsOverview(from: string, to: string): Promise<StatsOverview> {
-  return requestJson<StatsOverview>(`/api/stats/overview?${statsParams(from, to)}`);
-}
-
-export async function fetchArticlesOverTime(
-  from: string,
-  to: string
-): Promise<ArticlesOverTimePoint[]> {
-  const data = await requestJson<{ items: ArticlesOverTimePoint[] }>(
-    `/api/stats/articles-over-time?${statsParams(from, to)}`
-  );
-  return data.items;
-}
-
-export async function fetchSourcesVolume(from: string, to: string): Promise<SourceVolumePoint[]> {
-  const data = await requestJson<{ items: SourceVolumePoint[] }>(
-    `/api/stats/sources-volume?${statsParams(from, to)}`
-  );
-  return data.items;
-}
-
-export async function fetchArticleCounts(): Promise<ArticleCountsResult> {
-  return requestJson<ArticleCountsResult>('/api/stats/article-counts');
-}
-
-export async function fetchTriageMetrics(): Promise<TriageMetrics> {
-  return requestJson<TriageMetrics>('/api/stats/triage-metrics');
-}
-
-export async function fetchSourceQuality(): Promise<SourceQualityRow[]> {
-  const data = await requestJson<{ items: SourceQualityRow[] }>('/api/stats/source-quality');
-  return data.items;
-}
-
-export async function fetchCategoryMix(): Promise<CategoryMixPoint[]> {
-  const data = await requestJson<{ items: CategoryMixPoint[] }>('/api/stats/category-mix');
-  return data.items;
-}
-
-export async function fetchIngestedVsHandled(): Promise<IngestedVsHandledPoint[]> {
-  const data = await requestJson<{ items: IngestedVsHandledPoint[] }>(
-    '/api/stats/ingested-vs-handled'
-  );
-  return data.items;
-}
-
-export async function fetchIngestRuns(page = 1, perPage = 10): Promise<IngestRunPage> {
-  const params = new URLSearchParams({ page: String(page), per_page: String(perPage) });
-  return requestJson<IngestRunPage>(`/api/ingest/runs?${params}`);
-}
-
-export async function fetchIngestRunSources(runId: number): Promise<IngestRunSource[]> {
-  const data = await requestJson<{ items: IngestRunSource[] }>(`/api/ingest/runs/${runId}`);
-  return data.items;
-}
-
-export async function fetchLatestBriefing(): Promise<BriefingLatestResponse> {
-  return requestJson<BriefingLatestResponse>('/api/briefings/latest');
-}
-
-export async function createBriefing(focusPrompt?: string): Promise<BriefingCreateResponse> {
-  return requestJson<BriefingCreateResponse>('/api/briefings', {
-    method: 'POST',
-    ...(focusPrompt ? { body: JSON.stringify({ focus_prompt: focusPrompt }) } : {}),
-  });
-}
-
-export async function fetchBriefing(id: number): Promise<Briefing> {
-  return requestJson<Briefing>(`/api/briefings/${id}`);
-}
-
-export async function fetchBriefings(limit = 50, offset = 0): Promise<{ items: Briefing[] }> {
-  const params = new URLSearchParams({ limit: String(limit), offset: String(offset) });
-  return requestJson<{ items: Briefing[] }>(`/api/briefings?${params}`);
-}
-
-export async function generateBriefingPodcast(id: number): Promise<{ url: string }> {
-  return requestJson<{ url: string }>(`/api/briefings/${id}/podcast`, { method: 'POST' });
-}
-
-export interface PodcastFeedToken {
-  token: string;
-  url: string;
-}
-
-export async function fetchPodcastFeedToken(): Promise<PodcastFeedToken> {
-  return requestJson<PodcastFeedToken>('/api/briefings/podcast-feed-token');
-}
-
-export async function regeneratePodcastFeedToken(): Promise<PodcastFeedToken> {
-  return requestJson<PodcastFeedToken>('/api/briefings/podcast-feed-token/regenerate', {
-    method: 'POST',
-  });
-}
-
-export interface ChatMessage {
-  role: 'user' | 'assistant';
-  content: string;
-}
-
-export async function chatWithBriefing(
-  briefingId: number,
-  message: string,
-  history: ChatMessage[]
-): Promise<{ reply: string }> {
-  return requestJson<{ reply: string }>(`/api/briefings/${briefingId}/chat`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ message, history }),
-  });
-}
-
-// ── Auth ──────────────────────────────────────────────────────────────────────
-
-export interface AuthConfig {
-  provider: 'password' | 'keycloak';
-  keycloak_enabled: boolean;
-  login_url: string | null;
-  logout_url: string;
-  registration_url?: string | null;
-}
-
-export async function fetchAuthConfig(): Promise<AuthConfig> {
-  return requestJson<AuthConfig>('/api/auth/config');
-}
-
-export async function fetchMe(): Promise<User> {
-  return requestJson<User>('/api/auth/me');
-}
-
-export async function fetchAdminAnalytics(days = 30): Promise<AdminAnalytics> {
-  return requestJson<AdminAnalytics>(`/api/admin/analytics?days=${days}`);
-}
-
-export async function fetchAdminAiQuality(days = 30): Promise<AdminAiQuality> {
-  return requestJson<AdminAiQuality>(`/api/admin/ai/quality?days=${days}`);
-}
-
-export async function loginUser(username: string, password: string): Promise<User> {
-  return requestJson<User>('/api/auth/login', {
-    method: 'POST',
-    body: JSON.stringify({ username, password }),
-  });
-}
-
-export async function requestOtp(email: string): Promise<void> {
-  await requestJson<{ status: string }>('/api/auth/otp/request', {
-    method: 'POST',
-    body: JSON.stringify({ email }),
-  });
-}
-
-export async function loginWithOtp(email: string, otp: string): Promise<User> {
-  return requestJson<User>('/api/auth/otp/login', {
-    method: 'POST',
-    body: JSON.stringify({ email, otp }),
-  });
-}
-
-export interface GeneratedUser {
-  id: number | string | null;
-  username: string;
-  email?: string | null;
-  is_admin?: boolean;
-  password: string;
-  provider: 'keycloak' | 'password';
-  temporary?: boolean;
-  created_at?: string | null;
-}
-
-export async function fetchAdminUsers(): Promise<User[]> {
-  const data = await requestJson<{ items: User[] }>('/api/admin/users');
-  return data.items;
-}
-
-export async function generateAdminUser(
-  username: string,
-  options?: { email?: string | null; is_admin?: boolean }
-): Promise<GeneratedUser> {
-  return requestJson<GeneratedUser>('/api/admin/users/generate', {
-    method: 'POST',
-    body: JSON.stringify({
-      username,
-      email: options?.email ?? null,
-      is_admin: options?.is_admin ?? false,
-    }),
-  });
-}
-
-export async function deleteAdminUser(userId: number): Promise<void> {
-  await requestJson<{ status: string }>(`/api/admin/users/${userId}`, {
-    method: 'DELETE',
-  });
-}
-
-export async function logoutUser(): Promise<void> {
-  const config = await fetchAuthConfig().catch(() => null);
-  if (config?.provider === 'keycloak' && config.logout_url) {
-    window.location.assign(config.logout_url);
-    return;
-  }
-  await fetch('/api/auth/logout', { credentials: 'same-origin' });
-}
-
-export async function toggleSourceSubscription(
-  slug: string,
-  enabled: boolean
-): Promise<{ subscribed: boolean }> {
-  return requestJson<{ subscribed: boolean }>(`/api/sources/${slug}/enabled`, {
-    method: 'PATCH',
-    body: JSON.stringify({ enabled }),
-  });
-}
-
-/**
- * Recompute the current user's personalized recommendation scores on demand.
- * Returns how many articles were scored — zero means there's no interaction
- * history (starred/done/skipped) to learn from yet.
- */
-export async function recalculateMyRecommendations(): Promise<{ scored: number }> {
-  return requestJson<{ scored: number }>('/api/recommendations/recalculate-mine', {
-    method: 'POST',
-  });
-}
-
-// ── Notification settings & push subscriptions ────────────────────────────────
-
-export async function fetchNotificationSettings(): Promise<NotificationSettings> {
-  return requestJson<NotificationSettings>('/api/settings/notifications');
-}
-
-export async function updateNotificationSettings(
-  update: NotificationSettingsUpdate
-): Promise<Omit<NotificationSettings, 'vapid_public_key'>> {
-  return requestJson('/api/settings/notifications', {
-    method: 'PUT',
-    body: JSON.stringify(update),
-  });
-}
-
-export async function subscribePush(
-  payload: PushSubscribeRequest
-): Promise<{ subscribed: boolean }> {
-  return requestJson('/api/notifications/subscribe', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function unsubscribePush(endpoint?: string): Promise<{ unsubscribed: boolean }> {
-  return requestJson('/api/notifications/subscribe', {
-    method: 'DELETE',
-    body: endpoint ? JSON.stringify({ endpoint }) : undefined,
-  });
-}
-
-// ── Analytics (privacy) settings ───────────────────────────────────────────────
-
-export async function fetchAnalyticsSettings(): Promise<AnalyticsSettings> {
-  return requestJson<AnalyticsSettings>('/api/settings/analytics');
-}
-
-export async function updateAnalyticsSettings(
-  update: AnalyticsSettingsUpdate
-): Promise<AnalyticsSettings> {
-  return requestJson('/api/settings/analytics', {
-    method: 'PUT',
-    body: JSON.stringify(update),
-  });
-}
-
-// ─── In-platform sharing ──────────────────────────────────────────────────────
-
-export async function fetchShareableUsers(): Promise<ShareableUser[]> {
-  const data = await requestJson<{ items: ShareableUser[] }>('/api/users');
-  return data.items;
-}
-
-export async function shareArticle(
-  articleId: number,
-  toUserId: number,
-  note?: string
-): Promise<ReceivedShare> {
-  return requestJson<ReceivedShare>(`/api/articles/${articleId}/share`, {
-    method: 'POST',
-    body: JSON.stringify({ to_user_id: toUserId, note: note ?? null }),
-  });
-}
-
-export interface CreateShareAnnotationPayload {
-  highlighted_text: string;
-  offset_chars?: number;
-  note?: string | null;
-}
-
-export async function createShareAnnotation(
-  shareId: number,
-  payload: CreateShareAnnotationPayload
-): Promise<ShareAnnotation> {
-  return requestJson<ShareAnnotation>(`/api/shares/${shareId}/annotations`, {
-    method: 'POST',
-    body: JSON.stringify({
-      highlighted_text: payload.highlighted_text,
-      offset_chars: payload.offset_chars ?? 0,
-      note: payload.note ?? null,
-    }),
-  });
-}
-
-export async function fetchReceivedShares(): Promise<{
-  items: ReceivedShare[];
-  unread: number;
-}> {
-  return requestJson('/api/shares');
-}
-
-export async function fetchSharesUnreadCount(): Promise<number> {
-  const data = await requestJson<{ unread: number }>('/api/shares/unread_count');
-  return data.unread;
-}
-
-export async function fetchSentShares(): Promise<{ items: SentShare[] }> {
-  return requestJson('/api/shares/sent');
-}
-
-export async function revokeShare(shareId: number): Promise<void> {
-  await requestJson(`/api/shares/${shareId}/revoke`, { method: 'POST' });
-}
-
-// ── Reading Goals & Quizzes ───────────────────────────────────────────────────
-
-export async function fetchGoals(): Promise<ReadingGoal[]> {
-  const data = await requestJson<{ items: ReadingGoal[] }>('/api/goals');
-  return data.items;
-}
-
-export async function createGoal(description: string, keywords: string): Promise<ReadingGoal> {
-  return requestJson<ReadingGoal>('/api/goals', {
-    method: 'POST',
-    body: JSON.stringify({ description, keywords }),
-  });
-}
-
-export async function deleteGoal(goalId: number): Promise<void> {
-  await requestJson(`/api/goals/${goalId}`, { method: 'DELETE' });
-}
-
-export async function fetchQuizCandidates(): Promise<QuizCandidate[]> {
-  const data = await requestJson<{ candidates: QuizCandidate[] }>('/api/quizzes/candidates');
-  return data.candidates;
-}
-
-export async function fetchLatestQuiz(): Promise<Quiz | null> {
-  try {
-    return await requestJson<Quiz>('/api/quizzes/latest');
-  } catch (err) {
-    if (err instanceof HttpError && err.status === 404) return null;
-    throw err;
-  }
-}
-
-export async function fetchQuizHistory(): Promise<QuizHistoryItem[]> {
-  const data = await requestJson<{ items: QuizHistoryItem[] }>('/api/quizzes');
-  return data.items;
-}
-
-export async function generateQuiz(): Promise<Quiz> {
-  return requestJson<Quiz>('/api/quizzes/generate', { method: 'POST' });
-}
-
-export async function submitQuiz(quizId: number, answers: number[]): Promise<QuizResult> {
-  return requestJson<QuizResult>(`/api/quizzes/${quizId}/submit`, {
-    method: 'POST',
-    body: JSON.stringify({ answers }),
-  });
-}
-
-export async function markShareRead(shareId: number): Promise<void> {
-  await requestJson(`/api/shares/${shareId}/read`, { method: 'POST' });
-}
-
-export async function fetchShareDetail(shareId: number): Promise<ShareDetail> {
-  return requestJson<ShareDetail>(`/api/shares/${shareId}`);
-}
-
-export async function fetchShareMessages(shareId: number): Promise<{ items: ShareMessage[] }> {
-  return requestJson<{ items: ShareMessage[] }>(`/api/shares/${shareId}/messages`);
-}
-
-export async function postShareMessage(shareId: number, message: string): Promise<ShareMessage> {
-  return requestJson<ShareMessage>(`/api/shares/${shareId}/messages`, {
-    method: 'POST',
-    body: JSON.stringify({ message }),
-  });
-}
-
-export async function fetchTopicMap(): Promise<TopicMapResponse> {
-  return requestJson<TopicMapResponse>('/api/articles/topic-map');
-}
-
-export async function fetchAiWordCloud(days = 7): Promise<WordCloudResponse> {
-  return requestJson<WordCloudResponse>(`/api/ai-stats/word-cloud?days=${days}`);
-}
-
-export async function fetchAiEmbeddingMap(days = 7): Promise<EmbeddingMapResponse> {
-  return requestJson<EmbeddingMapResponse>(`/api/ai-stats/embedding-map?days=${days}`);
-}
-
-export async function fetchKnowledgeGraph(days = 7): Promise<KnowledgeGraphResponse> {
-  return requestJson<KnowledgeGraphResponse>(`/api/ai-stats/knowledge-graph?days=${days}`);
-}
-
-export async function fetchOnboardingStatus(): Promise<OnboardingStatus> {
-  return requestJson<OnboardingStatus>('/api/onboarding/status');
-}
-
-export async function fetchOnboardingInterests(): Promise<OnboardingInterest[]> {
-  return requestJson<OnboardingInterest[]>('/api/onboarding/interests');
-}
-
-export async function fetchOnboardingSourceRecommendations(
-  interestIds: string[]
-): Promise<OnboardingSourceRecommendation[]> {
-  return requestJson<OnboardingSourceRecommendation[]>('/api/onboarding/recommendations', {
-    method: 'POST',
-    body: JSON.stringify({ interest_ids: interestIds }),
-  });
-}
-
-export async function saveOnboardingInterests(
-  payload: SaveOnboardingInterestsRequest
-): Promise<void> {
-  await requestJson('/api/onboarding/interests', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function fetchPersonalizationNudges(): Promise<PersonalizationNudge[]> {
-  const data = await requestJson<{ items: PersonalizationNudge[] }>('/api/personalization/nudges');
-  return data.items;
-}
-
-export async function applyPersonalizationNudge(nudgeId: string): Promise<{ applied: boolean }> {
-  return requestJson('/api/personalization/nudges/apply', {
-    method: 'POST',
-    body: JSON.stringify({ nudge_id: nudgeId }),
-  });
-}
-
-export async function dismissPersonalizationNudge(
-  nudgeId: string,
-  cooldownDays = 7
-): Promise<{ dismissed: boolean }> {
-  return requestJson('/api/personalization/nudges/dismiss', {
-    method: 'POST',
-    body: JSON.stringify({ nudge_id: nudgeId, cooldown_days: cooldownDays }),
-  });
-}
-
-export async function fetchWatchlists(): Promise<AiWatchlist[]> {
-  const data = await requestJson<{ items: AiWatchlist[] }>('/api/watchlists');
-  return data.items;
-}
-
-export interface CreateWatchlistRequest {
-  label: string;
-  query: string;
-  threshold?: number;
-  enabled?: boolean;
-  notify_push?: boolean;
-}
-
-export async function createWatchlist(payload: CreateWatchlistRequest): Promise<AiWatchlist> {
-  return requestJson('/api/watchlists', {
-    method: 'POST',
-    body: JSON.stringify(payload),
-  });
-}
-
-export interface UpdateWatchlistRequest {
-  label?: string;
-  query?: string;
-  threshold?: number;
-  enabled?: boolean;
-  notify_push?: boolean;
-}
-
-export async function updateWatchlist(
-  watchlistId: number,
-  payload: UpdateWatchlistRequest
-): Promise<AiWatchlist> {
-  return requestJson(`/api/watchlists/${watchlistId}`, {
-    method: 'PATCH',
-    body: JSON.stringify(payload),
-  });
-}
-
-export async function deleteWatchlist(watchlistId: number): Promise<{ deleted: boolean }> {
-  return requestJson(`/api/watchlists/${watchlistId}`, { method: 'DELETE' });
-}
-
-export async function previewWatchlist(
-  query: string,
-  threshold?: number
-): Promise<AiWatchlistMatch[]> {
-  const data = await requestJson<{ items: AiWatchlistMatch[] }>('/api/watchlists/preview', {
-    method: 'POST',
-    body: JSON.stringify({ query, threshold }),
-  });
-  return data.items;
-}
-
-export async function fetchWatchlistNudges(): Promise<AiWatchlistNudge[]> {
-  const data = await requestJson<{ items: AiWatchlistNudge[] }>('/api/watchlists/nudges');
-  return data.items;
-}
-
-export async function downloadUserExport(): Promise<void> {
-  const data = await requestJson<unknown>('/api/users/me/export');
-  const blob = new Blob([JSON.stringify(data, null, 2)], { type: 'application/json' });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = `reading-archive-${new Date().toISOString().slice(0, 10)}.json`;
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-export async function deleteOwnAccount(confirmation: string): Promise<void> {
-  await requestJson<{ status: string }>('/api/users/me', {
-    method: 'DELETE',
-    body: JSON.stringify({ confirmation }),
-  });
-}
-
-// ── OPML import / export ──────────────────────────────────────────────────────
-
-export async function exportOpml(): Promise<void> {
-  const response = await fetch('/api/sources/export.opml', { credentials: 'same-origin' });
-  if (!response.ok) throw new Error(await readErrorMessage(response));
-  const blob = await response.blob();
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = 'subscriptions.opml';
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-export async function importOpml(file: File): Promise<OpmlImportResult> {
-  const form = new FormData();
-  form.append('file', file);
-  const response = await fetch('/api/sources/import', {
-    method: 'POST',
-    credentials: 'same-origin',
-    body: form,
-  });
-  if (!response.ok) throw new HttpError(response.status, await readErrorMessage(response));
-  return response.json() as Promise<OpmlImportResult>;
-}
